[zero] refactor legacy zero import path
