+++ colossalai/zero/legacy/gemini/__init__.py
@@ -1,3 +1,4 @@
+from .ophooks import register_ophooks_recursively
 from .stateful_tensor import StatefulTensor
 from .stateful_tensor_mgr import StatefulTensorMgr
 from .tensor_placement_policy import AutoTensorPlacementPolicy, CPUTensorPlacementPolicy, CUDATensorPlacementPolicy
@@ -10,4 +11,5 @@ __all__ = [
     'CUDATensorPlacementPolicy',
     'AutoTensorPlacementPolicy',
     'ZeroHook',
+    'register_ophooks_recursively',
 ]
