+++ colossalai/zero/legacy/shard_utils/tensor_shard_strategy.py
@@ -3,8 +3,8 @@ from typing import List, Optional
 import torch
 import torch.distributed as dist
 
-from colossalai.gemini.tensor_utils import colo_model_data_tensor_move_inline
 from colossalai.utils import get_current_device
+from colossalai.zero.legacy.gemini.tensor_utils import colo_model_data_tensor_move_inline
 from colossalai.zero.legacy.shard_utils import BaseShardStrategy
 from colossalai.zero.legacy.shard_utils.commons import get_shard
 from colossalai.zero.legacy.sharded_param.sharded_tensor import ShardedTensor
