+++ colossalai/zero/legacy/sharded_param/sharded_tensor.py
@@ -1,6 +1,6 @@
 import torch
 
-from colossalai.gemini.stateful_tensor import StatefulTensor, TensorState
+from colossalai.zero.legacy.gemini.stateful_tensor import StatefulTensor, TensorState
 
 
 class ShardedTensor(StatefulTensor):
